% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_intermediate}
\alias{create_intermediate}
\title{Create Intermediate}
\usage{
create_intermediate()
}
\value{
None
}
\description{
Creates a file in the current working directory used for program files later on
}
