% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cutadapt_run}
\alias{cutadapt_run}
\title{Cutadapt Run}
\usage{
cutadapt_run(cutadapt_path, cutadapt_data)
}
\arguments{
\item{cutadapt_path}{The path to the acutal cutadapt program}

\item{cutadapt_data}{The tibble from cutadapt tibble function}
}
\value{
None
}
\description{
To check for the user having cutadapt - if yes: run the cutadapt command
}
