% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{main_cutadapt_function}
\alias{main_cutadapt_function}
\title{Main Cutadapt Function}
\usage{
main_cutadapt_function(
  primer_path,
  metadata_path,
  fastq_path,
  intermediate_path,
  cutadapt_path
)
}
\arguments{
\item{primer_path}{a path to the csv that holds the primer info for this project.}

\item{metadata_path}{a path to the csv that holds the metadata info for this project.}

\item{fastq_path}{a path to the csv that holds the fastq for this project.}

\item{intermediate_path}{a path to the directory in which the intermediate folder should be created.}

\item{cutadapt_path}{a path to the cutadapt .exe file.}
}
\value{
A tibble that contains the forward and reverse primers with created complement, etc.
}
\description{
Funs all of the functions associated with running cutadapt. Users should not have to touch these functions.
}
