% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pre_primer_hit_data}
\alias{pre_primer_hit_data}
\title{Pre Primer Hit Data}
\usage{
pre_primer_hit_data(primer_data, fastq_data, intermediate_path)
}
\arguments{
\item{primer_data}{The primer data tibble created in prepare_primers function}

\item{fastq_data}{The tibble that the read_fastq function spits out}

\item{intermediate_path}{The path to the intermediate data file}
}
\value{
A data frame with the primer hits in each sample
}
\description{
To give a visualization (in a table) showing the different primer hits in each of the sequences for each sample
}
