% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{Prepare Metadata}
\usage{
prepare_metadata(metadata_path, primer_data)
}
\arguments{
\item{metadata_path}{Path to the metadata csv}

\item{primer_data}{The primer data tibble created in prepare_primers function}
}
\value{
A tibble containing the concatenated metadata and primer data
}
\description{
Read in the metadata from user and combine it with the primer data
}
