% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{primer_hit_plot}
\alias{primer_hit_plot}
\title{Primer Hit Plot}
\usage{
primer_hit_plot(primer_hits, fastq_data, metadata)
}
\arguments{
\item{primer_hits}{A tibble that shows the primer hits, produced by pre_primer_hit_data}

\item{fastq_data}{The tibble that the read_fastq function spits out}

\item{metadata}{The metadata table that prepare_metadata provides}
}
\value{
A plot that shows the primer hits
}
\description{
To create a plot showing the primer hits for a sanity check
}
