% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_fastq}
\alias{read_fastq}
\title{Read Fastq}
\usage{
read_fastq(fastq_path)
}
\arguments{
\item{fastq_path}{The path to the fastq files from user}
}
\value{
A tibble with the fastq file paths, the direction of the sequences, and names of sequences
}
\description{
Takes in the fastq files from the user and creates a tibble with the paths to future used files
}
